SubjectAcl resource added and applied. Extended wilcard permi…

+++ mica-rest/src/main/java/org/obiba/mica/dataset/rest/harmonization/DraftHarmonizationDatasetResource.java
@@ -34,6 +34,7 @@ import org.obiba.mica.dataset.domain.Dataset;
 import org.obiba.mica.dataset.domain.HarmonizationDataset;
 import org.obiba.mica.dataset.domain.HarmonizationDatasetState;
 import org.obiba.mica.dataset.service.HarmonizationDatasetService;
+import org.obiba.mica.security.rest.SubjectAclResource;
 import org.obiba.mica.web.model.Dtos;
 import org.obiba.mica.web.model.Mica;
 import org.obiba.opal.web.model.Magma;
@@ -67,20 +68,20 @@ public class DraftHarmonizationDatasetResource extends
 
   @GET
   public Mica.DatasetDto get() {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     return dtos.asDto(datasetService.findById(id), true);
   }
 
   @DELETE
   public void delete(){
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "DELETE", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "DELETE", id);
     datasetService.delete(id);
   }
 
   @PUT
   @Timed
   public Response update(Mica.DatasetDto datasetDto, @Context UriInfo uriInfo) {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "EDIT", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "EDIT", id);
     if(!datasetDto.hasId() || !datasetDto.getId().equals(id))
       throw new IllegalArgumentException("Not the expected dataset id");
     Dataset dataset = dtos.fromDto(datasetDto);
@@ -94,7 +95,7 @@ public class DraftHarmonizationDatasetResource extends
   @Path("/_index")
   @Timed
   public Response index() {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "EDIT", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "EDIT", id);
     datasetService.index(id);
     return Response.noContent().build();
   }
@@ -102,7 +103,7 @@ public class DraftHarmonizationDatasetResource extends
   @PUT
   @Path("/_publish")
   public Response publish() {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "PUBLISH", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "PUBLISH", id);
     datasetService.publish(id, true);
     return Response.noContent().build();
   }
@@ -110,7 +111,7 @@ public class DraftHarmonizationDatasetResource extends
   @DELETE
   @Path("/_publish")
   public Response unPublish() {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "PUBLISH", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "PUBLISH", id);
     datasetService.publish(id, false);
     return Response.noContent().build();
   }
@@ -118,14 +119,14 @@ public class DraftHarmonizationDatasetResource extends
   @GET
   @Path("/table")
   public Magma.TableDto getTable() {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     return datasetService.getTableDto(getDataset());
   }
 
   @GET
   @Path("/variables")
   public List<Mica.DatasetVariableDto> getVariables() {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     ImmutableList.Builder<Mica.DatasetVariableDto> builder = ImmutableList.builder();
     datasetService.getDatasetVariables(getDataset()).forEach(variable -> builder.add(dtos.asDto(variable)));
     return builder.build();
@@ -133,7 +134,7 @@ public class DraftHarmonizationDatasetResource extends
 
   @Path("/variable/{variable}")
   public DraftDataschemaDatasetVariableResource getVariable(@PathParam("variable") String variable) {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     DraftDataschemaDatasetVariableResource resource = applicationContext.getBean(DraftDataschemaDatasetVariableResource.class);
     resource.setDatasetId(id);
     resource.setVariableName(variable);
@@ -142,7 +143,7 @@ public class DraftHarmonizationDatasetResource extends
 
   @Path("/study/{study}/variable/{variable}")
   public DraftHarmonizedDatasetVariableResource getVariable(@PathParam("study") String studyId, @PathParam("variable") String variable) {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     DraftHarmonizedDatasetVariableResource resource = applicationContext.getBean(DraftHarmonizedDatasetVariableResource.class);
     resource.setDatasetId(id);
     resource.setVariableName(variable);
@@ -153,7 +154,7 @@ public class DraftHarmonizationDatasetResource extends
   @POST
   @Path("/facets")
   public List<Search.QueryResultDto> getFacets(Search.QueryTermsDto query) {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     ImmutableList.Builder<Search.QueryResultDto> builder = ImmutableList.builder();
     HarmonizationDataset dataset = getDataset();
     for(StudyTable table : dataset.getStudyTables()) {
@@ -166,7 +167,7 @@ public class DraftHarmonizationDatasetResource extends
   @Path("/_status")
   @Timed
   public Response updateStatus(@QueryParam("value") String status) {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "EDIT", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "EDIT", id);
     datasetService.updateStatus(id, RevisionStatus.valueOf(status.toUpperCase()));
 
     return Response.noContent().build();
@@ -175,8 +176,16 @@ public class DraftHarmonizationDatasetResource extends
   @GET
   @Path("/commit/{commitId}/view")
   public Mica.DatasetDto getFromCommit(@NotNull @PathParam("commitId") String commitId) throws IOException {
-    subjectAclService.isPermitted("/draft/harmonization-dataset", "VIEW", id);
+    subjectAclService.checkPermission("/draft/harmonization-dataset", "VIEW", id);
     return dtos.asDto(datasetService.getFromCommit(datasetService.findDraft(id), commitId), true);
+  }
+
+  @Path("/permissions")
+  public SubjectAclResource permissions() {
+    SubjectAclResource subjectAclResource = applicationContext.getBean(SubjectAclResource.class);
+    subjectAclResource.setResourceInstance("/draft/harmonization-dataset", id);
+    subjectAclResource.setFileResourceInstance("/draft/file", "/harmonization-dataset/" + id);
+    return subjectAclResource;
   }
 
   private HarmonizationDataset getDataset() {
