+++ mica-rest/src/main/java/org/obiba/mica/network/rest/DraftNetworkResource.java
@@ -32,6 +32,7 @@ import org.obiba.mica.network.NoSuchNetworkException;
 import org.obiba.mica.network.domain.Network;
 import org.obiba.mica.network.domain.NetworkState;
 import org.obiba.mica.network.service.NetworkService;
+import org.obiba.mica.security.rest.SubjectAclResource;
 import org.obiba.mica.web.model.Dtos;
 import org.obiba.mica.web.model.Mica;
 import org.springframework.context.ApplicationContext;
@@ -65,14 +66,14 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   @GET
   @Timed
   public Mica.NetworkDto get() {
-    subjectAclService.isPermitted("/draft/network", "VIEW", id);
+    subjectAclService.checkPermission("/draft/network", "VIEW", id);
     return dtos.asDto(networkService.findById(id), true);
   }
 
   @PUT
   @Timed
   public Response update(@SuppressWarnings("TypeMayBeWeakened") Mica.NetworkDto networkDto) {
-    subjectAclService.isPermitted("/draft/network", "EDIT", id);
+    subjectAclService.checkPermission("/draft/network", "EDIT", id);
     // ensure network exists
     networkService.findById(id);
 
@@ -85,7 +86,7 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   @Path("/_index")
   @Timed
   public Response index() {
-    subjectAclService.isPermitted("/draft/network", "EDIT", id);
+    subjectAclService.checkPermission("/draft/network", "EDIT", id);
     networkService.index(id);
     return Response.noContent().build();
   }
@@ -93,7 +94,7 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   @PUT
   @Path("/_publish")
   public Response publish() {
-    subjectAclService.isPermitted("/draft/network", "PUBLISH", id);
+    subjectAclService.checkPermission("/draft/network", "PUBLISH", id);
     networkService.publish(id);
     return Response.noContent().build();
   }
@@ -101,7 +102,7 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   @DELETE
   @Path("/_publish")
   public Response unPublish() {
-    subjectAclService.isPermitted("/draft/network", "PUBLISH", id);
+    subjectAclService.checkPermission("/draft/network", "PUBLISH", id);
     networkService.unPublish(id);
     return Response.noContent().build();
   }
@@ -109,7 +110,7 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   @DELETE
   @Timed
   public Response delete() {
-    subjectAclService.isPermitted("/draft/network", "DELETE", id);
+    subjectAclService.checkPermission("/draft/network", "DELETE", id);
     try {
       networkService.delete(id);
     } catch (NoSuchNetworkException e) {
@@ -122,7 +123,7 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   @Path("/_status")
   @Timed
   public Response toUnderReview(@QueryParam("value") String status) {
-    subjectAclService.isPermitted("/draft/network", "EDIT", id);
+    subjectAclService.checkPermission("/draft/network", "EDIT", id);
     networkService.updateStatus(id, RevisionStatus.valueOf(status.toUpperCase()));
 
     return Response.noContent().build();
@@ -146,6 +147,14 @@ public class DraftNetworkResource extends AbstractGitPersistableResource<Network
   public Mica.NetworkDto getFromCommit(@NotNull @PathParam("commitId") String commitId) throws IOException {
     subjectAclService.checkPermission("/draft/network", "VIEW", id);
     return dtos.asDto(networkService.getFromCommit(networkService.findDraft(id), commitId), true);
+  }
+
+  @Path("/permissions")
+  public SubjectAclResource permissions() {
+    SubjectAclResource subjectAclResource = applicationContext.getBean(SubjectAclResource.class);
+    subjectAclResource.setResourceInstance("/draft/network", id);
+    subjectAclResource.setFileResourceInstance("/draft/file", "/network/" + id);
+    return subjectAclResource;
   }
 
   @Override
